+++ software/source/clients/ios/README.md
@@ -1,6 +1,6 @@
 # iOS/Android Client
 
-***WORK IN PROGRESS***
+**_WORK IN PROGRESS_**
 
 This repository contains the source code for the 01 iOS/Android app. Work in progress, we will continue to improve this application to get it working properly.
 
@@ -9,10 +9,11 @@ Feel free to improve this and make a pull request!
 If you want to run it on your own, you will need to install Expo Go on your mobile device.
 
 ## Setup Instructions
+
 Follow the **[software setup steps](https://github.com/OpenInterpreter/01?tab=readme-ov-file#software)** in the main repo's README first before you read this
 
 ```shell
-cd software/source/clients/ios/react-native  # cd into `react-native`
+cd software/source/clients/mobile/react-native  # cd into `react-native`
 npm install                                  # install dependencies
 npx expo start                               # start local development server
 ```
@@ -20,6 +21,7 @@ npx expo start                               # start local development server
 In **Expo Go** select _Scan QR code_ to scan the QR code produced by the `npx expo start` command
 
 ## Using the App
+
 ```shell
 poetry run 01 --mobile                  # exposes QR code for 01 Light server
 ```
